BackEnd/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
distro==1.9.0
dnspython==2.6.1
fastapi==0.110.0
firebase-admin==6.5.0
google-api-core==2.18.0
google-api-python-client==2.123.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.6
install==1.3.5
msgpack==1.0.8
numpy==1.26.4
openai==1.14.3
opencv-python==4.9.0.80
outcome==1.3.0.post0
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
pymongo==4.6.2
pyparsing==3.1.2
PySocks==1.7.1
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.31.0
rsa==4.9
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.36.3
tqdm==4.66.2
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
wsproto==1.2.0